+++ anymotion_cli/commands/extract.py
@@ -1,3 +1,4 @@
+from pathlib import Path
 from typing import Optional
 
 import click
@@ -12,6 +13,7 @@ from ..state import State, pass_state
 from ..utils import color_id, get_client
 from .download import download_options
 from .draw import draw, draw_options
+from .upload import upload
 
 
 @click.group()
@@ -24,10 +26,18 @@ def cli() -> None:  # noqa: D103
     help_options_color="cyan",
     short_help="Extract keypoints from uploaded images or movies.",
 )
-@click.option("--movie-id", type=int)
-@click.option("--image-id", type=int)
+@click.option("--image-id", type=int, help="Uploaded image ID.")
+@click.option("--movie-id", type=int, help="Uploaded movie ID.")
 @click.option(
-    "-d", "--with-drawing", is_flag=True, help="Drawing with the extracted keypoints.",
+    "--path",
+    type=click.Path(exists=True, dir_okay=False),
+    help="The path of the movie or image file to extract.",
+)
+@click.option(
+    "-d",
+    "--with-drawing",
+    is_flag=True,
+    help="Drawing with the extracted keypoints.",
 )
 # TODO: remove download and draw option
 @draw_options
@@ -38,23 +48,34 @@ def cli() -> None:  # noqa: D103
 def extract(
     ctx: click.Context,
     state: State,
-    movie_id: Optional[int],
     image_id: Optional[int],
+    movie_id: Optional[int],
+    path: Path,
     with_drawing: bool,
     **kwargs,
 ) -> None:
-    """Extract keypoints from uploaded images or movies.
+    """Extract keypoints from image or movie.
 
-    Either "--image-id" or "--movie-id" is required.
+    Either '--image-id' or '--movie-id' or '--path' is required.
     """
-    if [movie_id, image_id].count(None) in [0, 2]:
-        raise click.UsageError("Either '--movie-id' or '--image-id' is required")
+    # Because path defaults to the current directory, not to None.
+    is_path = not path.is_dir()
+
+    if [bool(image_id), bool(movie_id), is_path].count(True) != 1:
+        raise click.UsageError(
+            "Either '--image-id' or '--movie-id' or '--path' is required"
+        )
 
     client = get_client(state)
+
+    result = None
+    if is_path:
+        result = ctx.invoke(upload, path=path)._asdict()
+        echo()
+    data = result or {"image_id": image_id, "movie_id": movie_id}
+
     try:
-        keypoint_id = client.extract_keypoint(
-            data={"image_id": image_id, "movie_id": movie_id}
-        )
+        keypoint_id = client.extract_keypoint(data=data)
         echo(f"Keypoint extraction started. (keypoint id: {color_id(keypoint_id)})")
 
         if state.use_spinner: