+++ anymotion_cli/core.py
@@ -1,7 +1,4 @@
 import click
-from click_repl import repl
-from prompt_toolkit.history import FileHistory
-from prompt_toolkit.styles import Style
 
 from . import __version__
 from .click_custom import CustomCommandCollection
@@ -18,7 +15,7 @@ from .commands.image import cli as image
 from .commands.keypoint import cli as keypoint
 from .commands.movie import cli as movie
 from .commands.upload import cli as upload
-from .config import get_app_dir
+from .interactive import run_interactive_mode
 from .options import profile_option
 from .state import State, pass_state
 
@@ -58,33 +55,6 @@ def cli(ctx: click.Context, state: State, interactive: bool) -> None:
 
     if ctx.invoked_subcommand is None:
         if interactive:
-            _run_interactive_mode(state)
+            run_interactive_mode(ctx, state)
         else:
             click.echo(cli.get_help(ctx))
-
-
-def _run_interactive_mode(state):
-    click.echo("Start interactive mode.")
-    click.echo(
-        "You can use the internal {help} command to explain usage.".format(
-            help=click.style(":help", fg="cyan")
-        )
-    )
-    click.echo()
-
-    style = Style.from_dict({"profile": "gray"})
-    message = [
-        ("class:cli_name", state.cli_name),
-        ("class:separator", " "),
-        ("class:profile", state.profile),
-        ("class:pound", "> "),
-    ]
-
-    repl(
-        click.get_current_context(),
-        prompt_kwargs={
-            "message": message,
-            "style": style,
-            "history": FileHistory(get_app_dir() / ".repl-history"),
-        },
-    )